Mod. updateStudent.routes, allow image durin sign

+++ routes/updateStudent.routes.js
@@ -31,17 +31,28 @@ updateStudentRouter.post('/:id', upload,  async (req, res) => {
 
         if (req.file) {
             new_image = req.file.filename;
-        
-            try {
-                const oldImagePath = path.join(__dirname, '../uploads/' + req.body.old_image);
-                fs.unlinkSync(oldImagePath);
-            } catch (error) {
-                if (error.code === 'ENOENT') {
-                    console.log('File does not exist:', error.message);
-                } else {
-                    console.error('Error deleting file:', error.message);
+            const oldImagePath = path.join(__dirname, '../uploads/' + req.body.old_image);
+
+            if (fs.existsSync(oldImagePath)) {
+                try{
+                    fs.unlinkSync(oldImagePath);
+                } catch (error) {
+                    console.log('Error deleting file:', error.message)
                 }
+            } else {
+                console.log('File does not exist:', oldImagePath)
             }
+        
+            // try {
+                
+            //     fs.unlinkSync(oldImagePath);
+            // } catch (error) {
+            //     if (error.code === 'ENOENT') {
+            //         console.log('File does not exist:', error.message);
+            //     } else {
+            //         console.error('Error deleting file:', error.message);
+            //     }
+            // }
         }
         
         const updatedDetails = {
